requirements.txt
fastapi
uvicorn
python-multipart
pypdf
sentence-transformers
faiss-cpu
pydantic
python-dotenv
pytest
httpx
